requirements.txt
langchain
langchain_community
langchain-core
langchain-classic
ipykernel
streamlit
python-dotenv
PyPDF2
pdf2image
pytesseract